requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
wandb>=0.15.0
jupyter>=1.0.0
tqdm>=4.64.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
streamlit>=1.28.0
